+++ projects/demo-maps/src/app/components/interpolation-settings/interpolation-settings.component.ts
@@ -2,6 +2,11 @@ import { Component, OnInit, Input } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { ColorRamp } from '@dlr-eoc/utils-maps';
 
+interface IinterpolationForm {
+  power: FormControl<number>;
+  smooth: FormControl<boolean>;
+  labels: FormControl<boolean>;
+}
 @Component({
   selector: 'app-interpolation-settings',
   templateUrl: './interpolation-settings.component.html',
@@ -11,14 +16,14 @@ export class InterpolationSettingsComponent implements OnInit {
 
   @Input() changeHandler: (power: number, smooth: boolean, labels: boolean) => void;
 
-  public interpolationForm: FormGroup;
+  public interpolationForm: FormGroup<IinterpolationForm>;
 
   constructor() {
-    this.interpolationForm = new FormGroup({
+    this.interpolationForm = new FormGroup<IinterpolationForm>({
       power: new FormControl(2, Validators.required),
       smooth: new FormControl(true, Validators.required),
       labels: new FormControl(false, Validators.required)
-  })
+    });
   }
 
   ngOnInit(): void {