+++ projects/demo-maps/src/app/route-components/bookmarks/bookmarks.component.ts
@@ -1,0 +1,36 @@
+import { Component, HostBinding, OnInit } from '@angular/core';
+import { Router, Routes } from '@angular/router';
+
+interface BookmarkEntry {
+  title: string;
+  route: string;
+  description?: string;
+  previewUrl?: string;
+}
+
+@Component({
+  selector: 'app-bookmarks',
+  templateUrl: './bookmarks.component.html',
+  styleUrls: ['./bookmarks.component.scss']
+})
+export class BookmarksComponent implements OnInit {
+  @HostBinding('class') class = 'content-container';
+  constructor(public router: Router) { }
+
+  routes: Routes;
+  availableBookmarks: BookmarkEntry[] = [];
+  ngOnInit(): void {
+    this.routes = this.router.config.filter(r => r.data);
+    this.availableBookmarks = this.routes.map(r => {
+      const bookmark: BookmarkEntry = {
+        title: r.data.title,
+        route: r.path,
+        description: r.data.description,
+        previewUrl: r.data.img
+      }
+      return bookmark;
+    });
+  }
+
+}
+